ai-services/ChatBot/requirements.txt
# Flask (original app)
flask==3.0.0
flask-cors==4.0.0

# FastAPI (new app)
fastapi==0.104.0
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6

# Vector Database & Embeddings
chromadb==0.4.17
langchain_chroma>=0.1.2
sentence-transformers==2.2.2

# LangChain & Ollama
langchain>=0.1.0
langchain_community>=0.1.0
langchain-ollama>=0.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.4.2
pypdf==3.16.4

# Production Server
gunicorn==21.2.0